Refactor contact_number fields in assessment-tool.js
This commit refactors the contact_number fields in the assessment-tool.js file. It adds two new fields, contact_number_2 and contact_number_3, to the list of fields that are retrieved in the getProblemsInEnvironment function. This change ensures that all contact numbers associated with a patient are included in the result, improving the completeness and accuracy of the data.